restore site in full
Earlier in the week, there was risks of a very possible stalking attack. Knowing the low IQ of some people who never learnt inspect element, it was safe to hide all the code that I didn't need.

Now that this situation is over, we're going to restore all of the original site in it's actual glory.